LayoutTests/http/tests/security-labels/cross-origin-iframe-expected.txt

CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL http://localhost:8000/security-labels/resources/cross-origin-iframe-iframe.html from frame with URL http://127.0.0.1:8000/security-labels/cross-origin-iframe.html. Domains, protocols and ports must match.

This tests to see if cross-origin frame interactions are still prohibited with security events.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS /cross-origin-iframe\.html$/.test(thisObj.location.href) is true
PASS event.destination.length == 0 is true
PASS crossFrameEvalFailed is true
PASS child location test
PASS child destination test
PASS successfullyParsed is true

TEST COMPLETE


--------
Frame: 'iframe'
--------
Helper iframe for testing cross-frame security events.
